requirements.txt
psycopg2
python-dotenv
# to install it run : pip install -r requirements.txt
